Refactor folder structure and sync namings
