+++ code/src/layouts/call-to-action/messages/en.json
@@ -1,0 +1,10 @@
+{
+  "callToAction.phones.caption": "Holiday Deals",
+  "callToAction.phones.title": "Up to\n30% off",
+  "callToAction.phones.description": "Selected Smartphone Brands",
+  "callToAction.phones.buttonLabel": "Shop",
+  "callToAction.computers.caption": "Just In",
+  "callToAction.computers.title": "Powerful\nmachines",
+  "callToAction.computers.description": "Top Computer Brands",
+  "callToAction.computers.buttonLabel": "Shop"
+}
