+++ code/src/layouts/call-to-action/messages/index.ts
@@ -1,0 +1,4 @@
+import en from "@/layouts/call-to-action/messages/en.json";
+import uk from "@/layouts/call-to-action/messages/uk.json";
+
+export default { en, uk };
