+++ code/src/layouts/call-to-action/messages/uk.json
@@ -1,0 +1,10 @@
+{
+  "callToAction.phones.caption": "Святкові пропозиції",
+  "callToAction.phones.title": "Знижки\nдо 30%",
+  "callToAction.phones.description": "Вибрані бренди смартфонів",
+  "callToAction.phones.buttonLabel": "Купити",
+  "callToAction.computers.caption": "Новинки",
+  "callToAction.computers.title": "Потужні\nмашини",
+  "callToAction.computers.description": "Топові бренди комп'ютерів",
+  "callToAction.computers.buttonLabel": "Купити"
+}
