+++ code/src/pages/home/HomePage.tsx
@@ -3,7 +3,7 @@ import IntroBanner from "@/layouts/intro-banner/IntroBanner";
 import PageWrapper from "@/layouts/page-wrapper/PageWrapper";
 import Subintro from "@/layouts/subintro/Subintro";
 import CategorySection from "@/layouts/category-section/CategorySection";
-import CallToActionSection from "@/pages/home/components/call-to-action/CallToActionSection";
+import CallToAction from "@/layouts/call-to-action/CallToAction";
 
 const HomePage = () => {
   return (
@@ -11,7 +11,7 @@ const HomePage = () => {
       <IntroBanner />
       <PageWrapper>
         <Subintro />
-        <CallToActionSection />
+        <CallToAction />
         <BestSellers />
         <CategorySection />
       </PageWrapper>
