--- code/src/pages/home/components/call-to-action/messages/en.json
@@ -1,10 +1,0 @@
-{
-  "callToActionSection.phones.caption": "Holiday Deals",
-  "callToActionSection.phones.title": "Up to\n30% off",
-  "callToActionSection.phones.description": "Selected Smartphone Brands",
-  "callToActionSection.phones.buttonLabel": "Shop",
-  "callToActionSection.computers.caption": "Just In",
-  "callToActionSection.computers.title": "Powerful\nmachines",
-  "callToActionSection.computers.description": "Top Computer Brands",
-  "callToActionSection.computers.buttonLabel": "Shop"
-}
