--- code/src/pages/home/components/call-to-action/messages/uk.json
@@ -1,10 +1,0 @@
-{
-  "callToActionSection.phones.caption": "Святкові пропозиції",
-  "callToActionSection.phones.title": "Знижки\nдо 30%",
-  "callToActionSection.phones.description": "Вибрані бренди смартфонів",
-  "callToActionSection.phones.buttonLabel": "Купити",
-  "callToActionSection.computers.caption": "Новинки",
-  "callToActionSection.computers.title": "Потужні\nмашини",
-  "callToActionSection.computers.description": "Топові бренди комп'ютерів",
-  "callToActionSection.computers.buttonLabel": "Купити"
-}
